fix: resolve TypeScript, ESLint, and Cargo clippy errors across codebase
Fix 49 TypeScript errors (mostly string|number type mismatches from
decimalLikeSchema in validation-utils), 7 ESLint errors, and Cargo clippy
warnings across multiple crates. Rename registry/registry.rs to
registry/provider_registry.rs to fix module_inception lint.

+++ crates/market-data/src/provider/yahoo/mod.rs
@@ -216,7 +216,7 @@ impl YahooProvider {
     ) -> Result<Quote, MarketDataError> {
         // Validate timestamp
         let timestamp: DateTime<Utc> = Utc
-            .timestamp_opt(yahoo_quote.timestamp as i64, 0)
+            .timestamp_opt(yahoo_quote.timestamp, 0)
             .single()
             .ok_or_else(|| MarketDataError::ValidationFailed {
                 message: format!("Invalid timestamp: {}", yahoo_quote.timestamp),
